SIT/requirements.txt
packageurl-python
pip_requirements_parser
PyYAML
requests
Jinja2
toml
pandas
cyclonedx-python-lib==7.5.1
spdx-tools